+++ Minigin/Transform.cpp
@@ -2,20 +2,20 @@
 #include <glm/gtc/matrix_transform.hpp>
 using namespace GameEngine;
 
-void GameEngine::Transform::SetPosition( float x,  float y,  float z)
+void GameEngine::Transform::SetPosition( int x,  int y,  int z)
 {
-	m_Position.x = x;
-	m_Position.y = y;
-	m_Position.z = z;
+	m_Position.x =static_cast<float>(x);
+	m_Position.y =static_cast<float>(y);
+	m_Position.z =static_cast<float>(z);
 	m_TranslateMatrix = glm::translate(glm::mat4(1.f), m_Position);
 }
 
-void GameEngine::Transform::Translate(float x, float y)
+void GameEngine::Transform::Translate(int x, int y)
 {
 	m_TranslateMatrix = glm::translate(m_TranslateMatrix, glm::vec3(x, y, 1.f));
 }
 
-void GameEngine::Transform::Translate(const glm::vec2& vec)
+void GameEngine::Transform::Translate(const glm::ivec2& vec)
 {
 	Translate(vec.x, vec.y);
 }
